fix user_flag type: uint32_t -> uint16_t across timeranger2 and queues
In yunetas V7 md2_record_ex_t.user_flag is uint16_t but several
wrapper functions still used uint32_t from V6. Fixed 10 locations
across 6 files:

- timeranger2.h: tranger2_write_user_flag, tranger2_set_user_flag
- timeranger2.c: same function implementations
- tr_queue.h/c: trq_set_hard_flag signature and __user_flag__ variable
- tr2_queue.h/c: tr2q_set_hard_flag signature and __user_flag__ variable

+++ kernel/c/timeranger2/src/tr_queue.c
@@ -445,7 +445,7 @@ PUBLIC int trq_check_pending_rowid(
     uint64_t __t__,
     uint64_t rowid
 ) {
-    uint32_t __user_flag__ = tranger2_read_user_flag(
+    uint16_t __user_flag__ = tranger2_read_user_flag(
         trq->tranger,
         trq->topic_name,
         "",
@@ -474,7 +474,7 @@ PUBLIC void trq_unload_msg(q_msg_t *msg, int32_t result)
     Put a hard flag in a message.
     You must flag a message after append it if you want recover it in the next open.
  ***************************************************************************/
-PUBLIC int trq_set_hard_flag(q_msg_t *msg, uint32_t hard_mark, BOOL set)
+PUBLIC int trq_set_hard_flag(q_msg_t *msg, uint16_t hard_mark, BOOL set)
 {
     return tranger2_set_user_flag(
         msg->trq->tranger,
