+++ yunos/c/mqtt_broker/src/tr2_queue.c
@@ -538,7 +538,7 @@ PUBLIC int tr2q_check_pending_rowid(
     uint64_t __t__,
     uint64_t rowid
 ) {
-    uint32_t __user_flag__ = tranger2_read_user_flag(
+    uint16_t __user_flag__ = tranger2_read_user_flag(
         trq->tranger,
         trq->topic_name,
         "",
@@ -557,7 +557,7 @@ PUBLIC int tr2q_check_pending_rowid(
     Put a hard flag in a message.
     You must flag a message after append it if you want recover it in the next open.
  ***************************************************************************/
-PUBLIC int tr2q_set_hard_flag(q2_msg_t *msg, uint32_t hard_mark, BOOL set)
+PUBLIC int tr2q_set_hard_flag(q2_msg_t *msg, uint16_t hard_mark, BOOL set)
 {
     return tranger2_set_user_flag(
         msg->trq->tranger,
